When initializing docpad, start with `docpad --offline init` before `run`.
This avoids the various explosions and error messages that happen if you
try to follow this tutorial in a directory where you have already set up
a git repository.

The correct fix is actually to have `docpad run` not explode if you
already have a `.git/` subdir that was not set up by DocPad itself from
a skeleton, but that's a larger change than I can do right now.